[#3857] switch between programs; split column.jsx
